refactor(ui): move configuration reload out of model.go
Watching files, comparing stamps, and reconciling a reloaded configuration is
one concern with its own failure rule: an invalid edit must leave the running
services alone.

Declarations move unchanged.
